Tel net wifi (#73)
* test heap size and delete dynamic values in webserver

* added html for redirection

* fixed escaped qutoes

* type cast

* no persenal info tracking

* typedasts

+++ adem/libraries/I2Cdev/I2Cdev.cpp
@@ -285,7 +285,7 @@ int8_t I2Cdev::readBytes(uint8_t devAddr, uint8_t regAddr, uint8_t length, uint8
             // I2C/TWI subsystem uses internal buffer that breaks with large data requests
             // so if user requests more than BUFFER_LENGTH bytes, we have to do it in
             // smaller chunks instead of all at once
-            for (uint8_t k = 0; k < length; k += min(length, BUFFER_LENGTH)) {
+            for (uint8_t k = 0; k < length; k += min((int)length, BUFFER_LENGTH)) {
                 for(int retry_cnt=0; retry_cnt < MAX_RETRY; retry_cnt++)
                 {
                     Wire.beginTransmission(devAddr);
